refactor code exec namespaces
