+++ python/packages/autogen-ext/src/autogen_ext/code_executors/docker/__init__.py
@@ -1,0 +1,3 @@
+from ._docker_code_executor import DockerCommandLineCodeExecutor
+
+__all__ = ["DockerCommandLineCodeExecutor"]
