+++ python/packages/autogen-ext/tests/code_executors/test_aca_dynamic_sessions.py
@@ -10,7 +10,7 @@ import pytest
 from anyio import open_file
 from autogen_core import CancellationToken
 from autogen_core.components.code_executor import CodeBlock
-from autogen_ext.code_executors import ACADynamicSessionsCodeExecutor
+from autogen_ext.code_executors.azure import ACADynamicSessionsCodeExecutor
 from azure.identity import DefaultAzureCredential
 
 UNIX_SHELLS = ["bash", "sh", "shell"]
