+++ python/packages/autogen-ext/tests/code_executors/test_aca_user_defined_functions.py
@@ -11,7 +11,7 @@ from autogen_core.components.code_executor import (
     FunctionWithRequirements,
     with_requirements,
 )
-from autogen_ext.code_executors import ACADynamicSessionsCodeExecutor
+from autogen_ext.code_executors.azure import ACADynamicSessionsCodeExecutor
 from azure.identity import DefaultAzureCredential
 
 ENVIRON_KEY_AZURE_POOL_ENDPOINT = "AZURE_POOL_ENDPOINT"
