+++ python/packages/autogen-magentic-one/examples/example_coder.py
@@ -11,7 +11,7 @@ from autogen_core import AgentId, AgentProxy
 from autogen_core.application import SingleThreadedAgentRuntime
 from autogen_core.application.logging import EVENT_LOGGER_NAME
 from autogen_core.components.code_executor import CodeBlock
-from autogen_ext.code_executors import DockerCommandLineCodeExecutor
+from autogen_ext.code_executors.docker import DockerCommandLineCodeExecutor
 from autogen_magentic_one.agents.coder import Coder, Executor
 from autogen_magentic_one.agents.orchestrator import RoundRobinOrchestrator
 from autogen_magentic_one.agents.user_proxy import UserProxy
